% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgmerge.R
\name{dev_compile}
\alias{dev_compile}
\title{Compile `.po` files to `.mo`}
\usage{
dev_compile(dir = ".", package = NULL, lazy = TRUE, verbose = TRUE)
}
\arguments{
\item{dir}{Path to package root directory.}

\item{package}{Name of package. If not supplied, read from `DESCRIPTION`.}

\item{lazy}{If `TRUE`, only `.mo` functions that are older than `.po`
files be updated}

\item{verbose}{If `TRUE`, print information as it goes.}
}
\description{
This function compiles the plain text `.po` files that translators work
to the binary `.mo` files that are installed with packages. In order for
translations to be accessible to R, you must re-compile every time the `.po`
files are changed.
}
