% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_extract.R
\name{po_extract}
\alias{po_extract}
\title{Extract messages for translation into a \code{.pot} file}
\usage{
po_extract(
  dir = ".",
  custom_translation_functions = list(),
  verbose = TRUE,
  style = NULL
)
}
\value{
The extracted messages as computed by \code{\link[=get_message_data]{get_message_data()}},
invisibly.
}
\description{
\code{po_extract()} scans your package for strings to be translated and
saves them in to \code{.pot} template file. You should never modify this
file by hand; instead modify the underlying source code and re-run
\code{po_extract()}.
}
