% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgmerge.R
\name{tr_add}
\alias{tr_add}
\title{Add a new translation}
\usage{
tr_add(lang, dir = ".")
}
\arguments{
\item{lang}{Language identifiers. These are typically two letters (e.g.
"en" = English, "fr" = French, "es" = Spanish, "zh" = Chinese), but
can include an additional suffix for languages that have regional
variations (e.g. "fr_CN" = French Canadian, "zh_CN" = simplified
characters as used in mainland China, "zh_TW" = traditional characters
as used in Taiwan.)}

\item{dir}{Path to package root.}
}
\description{
This creates `po/{lang}.po` containing the messages to be translated.
}
