% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_create.R, R/po_update.R
\name{po_create}
\alias{po_create}
\alias{po_update}
\title{Create a new \code{.po} file}
\usage{
po_create(languages, dir = ".", verbose = !is_testing())

po_update(dir = ".", lazy = TRUE, verbose = !is_testing())
}
\arguments{
\item{languages}{Language identifiers. These are typically two letters (e.g.
"en" = English, "fr" = French, "es" = Spanish, "zh" = Chinese), but
can include an additional suffix for languages that have regional
variations (e.g. "fr_CN" = French Canadian, "zh_CN" = simplified
characters as used in mainland China, "zh_TW" = traditional characters
as used in Taiwan.)}

\item{dir}{Character, default the present directory; a directory in which an
R package is stored.}

\item{verbose}{Logical, default \code{TRUE} (except during testing). Should
extra information about progress, etc. be reported?}

\item{lazy}{If \code{TRUE}, only \code{.po} files that are older than their
corresponding \code{.pot} file will be updated}
}
\description{
\code{po_create()} creates a new \verb{po/\{languages\}.po} containing the messages to be
translated. If a translation already exists, it'll be updated with any
changes to the \code{.pot} since it was last touched.

\code{po_update()} updates existing \code{.po} file after the \code{.pot} file has changed.
The translations for existing messages are preserved; new messages are added;
and translations for deleted message are marked as deprecated and moved to
the bottom of the file.
}
\seealso{
\code{\link[=po_update]{po_update()}} to update all \code{.po} files with changes from the
\code{.pot}.
}
