% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgmerge.R
\name{po_create}
\alias{po_create}
\alias{po_update}
\title{Create or update a \code{.po} file containing translations}
\usage{
po_create(languages, dir = ".", verbose = TRUE)

po_update(languages, dir = ".", verbose = TRUE)
}
\arguments{
\item{languages}{Language identifiers. These are typically two letters (e.g.
"en" = English, "fr" = French, "es" = Spanish, "zh" = Chinese), but
can include an additional suffix for languages that have regional
variations (e.g. "fr_CN" = French Canadian, "zh_CN" = simplified
characters as used in mainland China, "zh_TW" = traditional characters
as used in Taiwan.)}

\item{dir}{Path to package root.}

\item{verbose}{If \code{TRUE}, explain what's happening.}
}
\description{
\itemize{
\item \code{po_create()} creates a new \verb{po/\{languages\}.po} containing the messages to be
translated.
\item \code{po_update()} updates an existing \code{.po} file after the messages in a
package have changed. The translations for existing messages are preserved;
new messages are added; and translations for deleted message are marked
as deprecated and moved to the bottom of the file.
}
}
