% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_extract.R
\name{po_extract}
\alias{po_extract}
\title{Extract messages for translation into a `.pot` file}
\usage{
po_extract(
  dir = ".",
  custom_translation_functions = list(),
  verbose = TRUE,
  style = c("base", "explicit")
)
}
\value{
The extracted messages as computed by [get_message_data()],
  invisibly.
}
\description{
`po_extract()` scans your package for strings to be translated and
saves them in to `.pot` template file. You should never modify this
file by hand; instead modify the underlying source code and re-run
`po_extract()`.
}
