% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potools-package.R
\docType{package}
\name{potools-package}
\alias{potools}
\alias{potools-package}
\title{potools: Tools for Internationalization and Portability in R Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Translating messages in R packages is managed using the po top-level directory and the 'gettext' program. This package provides some helper functions for building this support in R packages, e.g. common validation & I/O tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MichaelChirico/potools}
  \item \url{https://michaelchirico.github.io/potools/}
  \item Report bugs at \url{https://github.com/MichaelChirico/potools/issues}
}

}
\author{
\strong{Maintainer}: Michael Chirico \email{MichaelChirico4@gmail.com}

Authors:
\itemize{
  \item Hadley Wickham
}

}
\keyword{internal}
