% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_package.R
\name{translate_package}
\alias{translate_package}
\title{Interactively provide translations for a package's messages}
\usage{
translate_package(
  dir = ".",
  languages,
  diagnostics = list(check_cracked_messages, check_untranslated_cat,
    check_untranslated_src),
  custom_translation_functions = list(R = NULL, src = NULL),
  max_translations = Inf,
  use_base_rules = package \%chin\% .potools$base_package_names,
  copyright = NULL,
  bugs = "",
  verbose = !is_testing()
)
}
\arguments{
\item{dir}{Character, default the present directory; a directory in which an
R package is stored.}

\item{languages}{Character vector; locale codes to which to translate. See
Details.}

\item{diagnostics}{A \code{list} of diagnostic functions to be run on the
package's message data. See Details.}

\item{custom_translation_functions}{A \code{list} with either/both of two
components, \code{R} and \code{src}, together governing how to extract any
non-standard strings from the package. See Details.}

\item{max_translations}{Numeric; used for setting a cap on the number of
translations to be done for each language. Defaults to \code{Inf}, meaning
all messages in the package.}

\item{use_base_rules}{Logical; Should internal behavior match base behavior
as strictly as possible? \code{TRUE} if being run on a base package (i.e.,
\code{base} or one of the default packages like \code{utils},
\code{graphics}, etc.). See Details.}

\item{copyright}{Character; passed on to \code{\link{write_po_file}}.}

\item{bugs}{Character; passed on to \code{\link{write_po_file}}.}

\item{verbose}{Logical, default \code{TRUE} (except during testing). Should
extra information about progress, etc. be reported?}
}
\value{
This function returns nothing invisibly. As a side effect, a
\file{.pot} file is written to the package's \file{po} directory (updated if
one does not yet exist, or created from scratch otherwise), and a \file{.po}
file is written in the same directory for each element of \code{languages}.
}
\description{
This function handles the "grunt work" of building and updating translation
libraries. In addition to providing a friendly interface for supplying
translations, some internal logic is built to help make your package more
translation-friendly.
}
\details{

}
\examples{

pkg <- system.file('pkg', package = 'potools')
# copy to a temporary location to be able to read/write/update below
tmp_pkg <- file.path(tempdir(), "pkg")
dir.create(tmp_pkg)
file.copy(pkg, dirname(tmp_pkg), recursive = TRUE)

# run translate_package() without any languages
# this will generate a .pot template file and en@quot translations (in UTF-8 locales)
# we can also pass empty 'diagnostics' to skip the diagnostic step
# (skip if gettext isn't available to avoid an error)
if (isTRUE(check_potools_sys_reqs)) {
  translate_package(tmp_pkg, diagnostics = NULL)
}

\dontrun{
# launches the interactive translation dialog for translations into Estonian:
translate_package(tmp_pkg, "et_EE", diagnostics = NULL, verbose = TRUE)
}

# cleanup
unlink(tmp_pkg, recursive = TRUE)
rm(pkg, tmp_pkg)

}
\references{
\url{https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Internationalization}
\cr
\url{https://cran.r-project.org/doc/manuals/r-release/R-admin.html#Internationalization}
\cr
\url{https://cran.r-project.org/doc/manuals/r-release/R-ints.html#Internationalization-in-the-R-sources}
\cr \url{https://developer.r-project.org/Translations30.html} \cr
\url{https://www.isi-web.org/publications/glossary-of-statistical-terms} \cr
\url{https://www.gnu.org/software/gettext/} \cr
\url{https://www.gnu.org/software/gettext/manual/html_node/Usual-Language-Codes.html#Usual-Language-Codes}
\cr
\url{https://www.gnu.org/software/gettext/manual/html_node/Country-Codes.html#Country-Codes}
\cr \url{https://www.stats.ox.ac.uk/pub/Rtools/goodies/gettext-tools.zip}
\cr \url{https://saimana.com/list-of-country-locale-code/}
}
\seealso{
\code{\link{get_message_data}}, \code{\link{write_po_file}},
\code{\link[tools]{xgettext}}, \code{\link[tools]{update_pkg_po}},
\code{\link[tools]{checkPoFile}}, \code{\link[base]{gettext}}
}
\author{
Michael Chirico
}
